[#183] Refactor: 대시보드 Navbar Profile 부분 커스텀 훅으로 정리
